+++ CHANGELOG.md
@@ -12,6 +12,9 @@
     - `:telegram_bot` rspec tag is replaced with `telegram_bot: :rails`.
 - __Breaking change__. Use bang-methods as actions for commands.
   This prevents calling context contextual actions and payload specific actions with commands.
+- __Breaking change__. Drop `.context_handler`, `.context_to_action!` methods.
+  Use pass action name directly to `#save_context`.
+  It's the same as `.context_to_action!` is enabled by default.
 
 # 0.13.1